Fixed damned bug.
What a delightful one. What it boils down to is that gcc initializes your member variables (ints, in particular) for you, to
-1. I had a typo in an if statement comparing against mIndex (my member variable) rather than aIndex (my parameter).

Further, there was an error with how I was handling End values. My default starting index was -1, the same as my End flag index, creating a meaning collision. Poop. So, I fixed that. Who knows what the repercussions would have been otherwise.

That kind of thing is just bad design, and I'll need to go back and fix it in the future.