Added variable read length support to Pyper.
Pyper slice iterator now distinguishes between reads that actually overlap with the target segment and ones that merely start at an index that should include overlaps if the reads are long enough.

Todo (Maybe):
Remove Caper "intersection" functionality from the engine. Rather, a Pyper style slice iterator will be implemented as a wrapper over the engine, but in C++.